Rewrite rate limit middleware to match new middleware signature
